refactor: change structure, delete unnecessary code
