+++ src/components/Header.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import "../css/header-footer.css";
 import { Link } from "react-router-dom";
 import { useDispatch } from "react-redux";
-import { setShowResults } from "../store/letsVoteBlockSlice";
+import { setShowResults } from "../store/votingPageSlice";
 
 export default function Header() {
   const dispatch = useDispatch();
