+++ src/components/QuestionForm.tsx
@@ -20,6 +20,7 @@ export default function QuestionForm() {
   const handleOnSubmit = (data: IQuestionInput) => {
     const newQuestion = data.questionInput;
     dispatch(setQuestion(newQuestion));
+    console.log(question, data.questionInput);
   };
 
   return (
