+++ src/components/servise.tsx
@@ -2,6 +2,37 @@ import { IOption } from "../store/optionsFormSlice";
 
 const VoteBlockHeight = 350;
 
+export default function createVotingBlock(
+  optionsArr: IOption[],
+  toShowPercentage: boolean,
+  toShowResult: boolean
+) {
+  const clicks = getTotalClicks(optionsArr);
+
+  const results = optionsArr.map((el, index) => {
+    const height = getHight(clicks, el.clicked);
+    return (
+      <div
+        className="result"
+        key={index}
+        style={{
+          width: getWidth(optionsArr),
+          background: el.color,
+          height: toShowResult ? height + "px" : "5px",
+        }}
+      >
+        {toShowPercentage && (
+          <span className="percentage">
+            {toShowResult ? getPercentage(height) + "%" : "0%"}
+          </span>
+        )}
+      </div>
+    );
+  });
+
+  return results;
+}
+
 export function getRandomColor() {
   var letters = "0123456789ABCDEF";
   var color = "#";
